figma-transformer: extract vector SVG renderer
